requirements.txt
Flask==2.3.2
Flask-Cors==3.0.10
pymongo==4.5.0
Flask-PyMongo==2.3.0
python-dotenv==1.0.0
marshmallow==3.19.0
Flask-Logging==0.1.3
Flask-RESTful==0.3.9
Flask-JWT-Extended==4.4.4
dnspython==2.3.0
mongoengine==0.27.0
pycodestyle==2.12.1
pre-commit==3.8.0
Flask-ReDoc==0.2.1
flasgger==0.9.7.1
